move styling to component directories
