+++ src/frontend/src/components/Backdrop/backdrop.css
@@ -1,11 +1,13 @@
+.backdrop {
+  position: fixed;
+  width: 100%;
+  height: 100vh;
+  background: rgba(255, 255, 255, 0.5);
+  z-index: 100;
+  top: 0;
+  left: 0;
+  animation: transform 0.3s ease-out;
+}
 .backdrop.open {
   display: none;
-}
-
-.text-area {
-  width: 80vw;
-  background: rgba(255, 255, 255, 0.7);
-  margin: 3rem auto 0;
-  font-size: 1.7rem;
-  padding: rem 10rem;
 }
