+++ src/frontend/src/components/Layout/Layout.js
@@ -1,16 +1,16 @@
 import React, { Component } from 'react';
 
-import Header from './Header';
-import SideDrawer from './SideDrawer';
-import Backdrop from './Backdrop';
-import TextArea from './TextArea';
-import Logo from './shared/Logo';
+import Header from '../Header/Header';
+import SideDrawer from '../SideDrawer/SideDrawer';
+import Backdrop from '../Backdrop';
+import TextArea from '../TextArea/TextArea';
+import Logo from '../Logo/Logo';
 
-import logo from '../images/logo.svg';
+import logo from '../../images/logo.svg';
 
-import data from '../../../../dummyData.json';
+import data from '../../../../../dummyData.json';
 
-import '../css/style.css';
+import './layout.css';
 
 const items = [
   { link: '#', text: 'Home' },
