+++ src/frontend/src/components/SideDrawer/SideDrawer.jsx
@@ -1,7 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import List from './shared/List/List';
-import Logo from './shared/Logo';
+import List from '../shared/List/List.jsx';
+import Logo from '../Logo/Logo.jsx';
+
+import './sideDrawer.css';
 
 const SideDrawer = ({ className, items, show, logo }) => (
   <nav className={`${className} ${show ? 'open' : ''}`}>
