+++ src/frontend/src/components/SideDrawer/sideDrawer.css
@@ -1,0 +1,42 @@
+.sidedrawer {
+  background-color: rgba(36, 36, 36, 0.9);
+  width: 20vw;
+  height: 100vh;
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 200;
+  max-width: 400px;
+  padding: 10px;
+  transform: translateX(-100%);
+  transition: transform 0.3s ease-out;
+  padding-top: 5rem;
+}
+.sidedrawer__logo {
+  text-align: center;
+  max-height: 20rem;
+}
+.sidedrawer__logo img {
+  height: 10rem;
+}
+.sidedrawer.open {
+  transform: translateX(0);
+}
+.sidedrawer__navigation-items {
+  height: 100%;
+  list-style: none;
+  display: flex;
+  flex-direction: column;
+}
+.sidedrawer__navigation-item {
+  margin: 5px 5px;
+  text-align: justify;
+}
+.sidedrawer__navigation-item a {
+  display: inline-block;
+  width: 100%;
+  color: #cccccc;
+  font-size: 1.7rem;
+  text-decoration: none;
+  border-bottom: 1px solid rgba(204, 204, 204, 0.2);
+}
