+++ src/frontend/src/components/TextArea/TextArea.jsx
@@ -1,7 +1,9 @@
 import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 
-import Post from './shared/Post';
+import Post from '../Post/Post.jsx';
+
+import './textArea.css';
 
 const TextArea = ({ className, posts }) => {
   return (
