+++ src/frontend/src/pages/index.css
@@ -1,0 +1,33 @@
+*,
+*::after,
+*::before {
+  margin: 0;
+  padding: 0;
+  box-sizing: inherit;
+}
+
+html {
+  /*assuming browser font default is 16px*/
+  font-size: 62.5%;
+  /*translates to 10px -> root font*/
+}
+
+/* font properties are stated here
+because they will get inherited by
+child components */
+body {
+  box-sizing: border-box;
+  height: 100vh;
+  background: url('../images/background.jpg') no-repeat center fixed;
+  -webkit-background-size: cover;
+  -moz-background-size: cover;
+  -o-background-size: cover;
+}
+
+html,
+body {
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+}
