+++ src/frontend/src/pages/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import Layout from '../components/layout';
+import Layout from '../components/Layout/Layout';
+import './index.css';
 
 export default function IndexPage() {
   return (
